requirements.txt
dnspython==2.7.0
pymongo==4.13.2
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dotenv==1.1.1
shiboken6==6.9.1
